# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Reciclado de Alimentos - API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register/merchant:
    post:
      operationId: registerMerchant
      tags: [auth]
      summary: Register a merchant account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterMerchantBody"
      responses:
        "201":
          description: Merchant registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register/buyer:
    post:
      operationId: registerBuyer
      tags: [auth]
      summary: Register a buyer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBuyerBody"
      responses:
        "201":
          description: Buyer registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /merchant/products:
    get:
      operationId: listMerchantProducts
      tags: [merchants]
      summary: List products for current merchant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Merchant products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /merchant/profile:
    get:
      operationId: getMerchantProfile
      tags: [merchants]
      summary: Get current merchant profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Merchant profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateMerchantProfile
      tags: [merchants]
      summary: Update merchant profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMerchantProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantProfile"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /merchant/orders:
    get:
      operationId: listMerchantOrders
      tags: [merchants]
      summary: List orders for current merchant
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Merchant orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /merchant/orders/validate:
    post:
      operationId: validateOrder
      tags: [merchants]
      summary: Validate and confirm order delivery by QR code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateOrderBody"
      responses:
        "200":
          description: Order validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid or already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /merchant/stats:
    get:
      operationId: getMerchantStats
      tags: [merchants]
      summary: Get merchant dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Merchant stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantStats"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
        merchantProfile:
          oneOf:
            - $ref: "#/components/schemas/MerchantProfile"
            - type: "null"
        buyerProfile:
          oneOf:
            - $ref: "#/components/schemas/BuyerProfile"
            - type: "null"
      required:
        - id
        - email
        - role
    MerchantProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessName:
          type: string
        legalName:
          type: ["string", "null"]
        cuit:
          type: ["string", "null"]
        category:
          type: string
        description:
          type: ["string", "null"]
        phone:
          type: string
        addressLine:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        pickupHours:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isOpen:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - businessName
        - category
        - phone
        - addressLine
        - city
        - province
        - isVerified
        - isOpen
        - createdAt
    BuyerProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        addressLine:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - firstName
        - lastName
        - phone
        - createdAt
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        icon:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - id
        - name
        - slug
        - isActive
    Product:
      type: object
      properties:
        id:
          type: string
        merchantId:
          type: string
        categoryId:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        originalPrice:
          type: number
        salePrice:
          type: number
        discountPercent:
          type: number
        quantityAvailable:
          type: integer
        unit:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        pickupDeadline:
          type: ["string", "null"]
        pickupAddress:
          type: ["string", "null"]
        pickupLatitude:
          type: ["number", "null"]
        pickupLongitude:
          type: ["number", "null"]
        status:
          type: string
        isFeatured:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        merchant:
          $ref: "#/components/schemas/MerchantProfile"
        category:
          $ref: "#/components/schemas/Category"
      required:
        - id
        - merchantId
        - categoryId
        - name
        - slug
        - originalPrice
        - salePrice
        - discountPercent
        - quantityAvailable
        - status
        - isFeatured
        - createdAt
        - merchant
        - category
...
    Order:
      type: object
      properties:
        id:
          type: string
        buyerId:
          type: string
        productId:
          type: string
        merchantId:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
        buyerFee:
          type: number
        merchantFee:
          type: number
        total:
          type: number
        paymentStatus:
          type: string
        deliveryStatus:
          type: string
        qrCode:
          type: string
        pickupCode:
          type: string
        deliveredAt:
          type: ["string", "null"]
        cancelledAt:
          type: ["string", "null"]
        createdAt:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        merchant:
          $ref: "#/components/schemas/MerchantProfile"
        buyer:
          $ref: "#/components/schemas/BuyerProfile"
      required:
        - id
        - buyerId
        - productId
        - merchantId
        - quantity
        - unitPrice
        - subtotal
        - buyerFee
        - merchantFee
        - total
        - paymentStatus
        - deliveryStatus
        - qrCode
        - pickupCode
        - createdAt
        - product
        - merchant
        - buyer
    MerchantStats:
      type: object
      properties:
        totalProducts:
          type: integer
        activeProducts:
          type: integer
        soldOutProducts:
          type: integer
        totalOrdersToday:
          type: integer
        totalOrdersMonth:
          type: integer
        revenueToday:
          type: number
        revenueMonth:
          type: number
        pendingPickup:
          type: integer
        delivered:
          type: integer
        platformCommission:
          type: number
      required:
        - totalProducts
        - activeProducts
        - soldOutProducts
        - totalOrdersToday
        - totalOrdersMonth
        - revenueToday
        - revenueMonth
        - pendingPickup
        - delivered
        - platformCommission
...
    RegisterMerchantBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        businessName:
          type: string
        legalName:
          type: ["string", "null"]
        cuit:
          type: ["string", "null"]
        category:
          type: string
        phone:
          type: string
        addressLine:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        pickupHours:
          type: ["string", "null"]
      required:
        - email
        - password
        - businessName
        - category
        - phone
        - addressLine
        - city
        - province
    RegisterBuyerBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        addressLine:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
      required:
        - email
        - password
        - firstName
        - lastName
        - phone
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
...
    UpdateMerchantProfileBody:
      type: object
      properties:
        businessName:
          type: ["string", "null"]
        legalName:
          type: ["string", "null"]
        cuit:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        addressLine:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        pickupHours:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        isOpen:
          type: ["boolean", "null"]
...
    ValidateOrderBody:
      type: object
      properties:
        pickupCode:
          type: string
      required:
        - pickupCode