# lib/api-zod/src/generated/types/merchantStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reciclado de Alimentos - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MerchantStats {
  totalProducts: number;
  activeProducts: number;
  soldOutProducts: number;
  totalOrdersToday: number;
  totalOrdersMonth: number;
  revenueToday: number;
  revenueMonth: number;
  pendingPickup: number;
  delivered: number;
  platformCommission: number;
}