# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reciclado de Alimentos - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BuyerProfile } from "./buyerProfile";
import type { MerchantProfile } from "./merchantProfile";
import type { Product } from "./product";

export interface Order {
  id: string;
  buyerId: string;
  productId: string;
  merchantId: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
  buyerFee: number;
  merchantFee: number;
  total: number;
  paymentStatus: string;
  deliveryStatus: string;
  qrCode: string;
  pickupCode: string;
  /** @nullable */
  deliveredAt?: string | null;
  /** @nullable */
  cancelledAt?: string | null;
  createdAt: string;
  product: Product;
  merchant: MerchantProfile;
  buyer: BuyerProfile;
}