# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reciclado de Alimentos - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";
import type { MerchantProfile } from "./merchantProfile";

export interface Product {
  id: string;
  merchantId: string;
  categoryId: string;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  originalPrice: number;
  salePrice: number;
  discountPercent: number;
  quantityAvailable: number;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  expiryDate?: string | null;
  /** @nullable */
  pickupDeadline?: string | null;
  /** @nullable */
  pickupAddress?: string | null;
  /** @nullable */
  pickupLatitude?: number | null;
  /** @nullable */
  pickupLongitude?: number | null;
  status: string;
  isFeatured: boolean;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  merchant: MerchantProfile;
  category: Category;
}