# lib/api-zod/src/generated/types/updateMerchantProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reciclado de Alimentos - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMerchantProfileBody {
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  legalName?: string | null;
  /** @nullable */
  cuit?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  addressLine?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  province?: string | null;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  pickupHours?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  isOpen?: boolean | null;
}